Move ValueSnapshotter and friends to snapshots project
