+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/rules/InputPropertyTaskStateChanges.java
@@ -19,7 +19,7 @@ package org.gradle.api.internal.changedetection.rules;
 import org.gradle.api.NonNullApi;
 import org.gradle.api.Task;
 import org.gradle.api.internal.changedetection.state.TaskExecution;
-import org.gradle.api.internal.changedetection.state.ValueSnapshot;
+import org.gradle.internal.snapshot.ValueSnapshot;
 
 import java.util.SortedMap;
