+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/rules/InputPropertyValueTaskStateChanges.java
@@ -21,8 +21,8 @@ import com.google.common.collect.ImmutableSortedMap;
 import org.gradle.api.internal.TaskInternal;
 import org.gradle.api.internal.changedetection.state.ImplementationSnapshot;
 import org.gradle.api.internal.changedetection.state.TaskExecution;
-import org.gradle.api.internal.changedetection.state.ValueSnapshot;
 import org.gradle.internal.changes.TaskStateChangeVisitor;
+import org.gradle.internal.snapshot.ValueSnapshot;
 
 import java.util.Map;
