+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/state/SnapshotSerializer.java
@@ -22,6 +22,7 @@ import org.gradle.internal.serialize.Decoder;
 import org.gradle.internal.serialize.Encoder;
 import org.gradle.internal.serialize.HashCodeSerializer;
 import org.gradle.internal.serialize.Serializer;
+import org.gradle.internal.snapshot.ValueSnapshot;
 
 import java.util.Map;
