+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/state/TaskExecutionFingerprintSerializer.java
@@ -26,6 +26,7 @@ import org.gradle.internal.serialize.AbstractSerializer;
 import org.gradle.internal.serialize.Decoder;
 import org.gradle.internal.serialize.Encoder;
 import org.gradle.internal.serialize.Serializer;
+import org.gradle.internal.snapshot.ValueSnapshot;
 
 import java.util.Map;
