+++ subprojects/snapshots/src/main/java/org/gradle/internal/snapshot/ValueSnapshottable.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 the original author or authors.
+ * Copyright 2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,17 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.gradle.internal.snapshot;
 
-package org.gradle.api.internal.changedetection.state;
-
-import org.gradle.internal.hash.Hashable;
-
-/**
- * An immutable snapshot of the state of some Java object or object graph.
- */
-public interface ValueSnapshot extends Hashable {
-    /**
-     * Takes a snapshot of the given value, using this as a candidate snapshot. If the value is the same as the value represented by this snapshot, this snapshot _must_ be returned.
-     */
-    ValueSnapshot snapshot(Object value, ValueSnapshotter snapshotter);
+public interface ValueSnapshottable extends Snapshottable<ValueSnapshot> {
 }
